Create Huffman.java
This program takes a sentence from the user and compresses it using the Huffman compression algorithm. It requires a Node and Tree class which I've uploaded in this repository. The program outputs the compression ratio achieved. I plan to adapt this code later to read in a text file and save the output in a new file similar to a .zip.